do not load query executor during dmmf command
