+++ libs/datamodel/core/src/configuration/datasource.rs
@@ -1,4 +1,4 @@
-use datamodel_connector::Connector;
+use datamodel_connector::{Connector, ConnectorCapabilities};
 use serde::Serialize;
 
 /// a `datasource` from the prisma schema.
@@ -19,6 +19,11 @@ pub struct Datasource {
 impl Datasource {
     pub fn url(&self) -> &StringFromEnvVar {
         &self.url
+    }
+
+    pub fn capabilities(&self) -> ConnectorCapabilities {
+        let capabilities = self.active_connector.capabilities().clone();
+        ConnectorCapabilities::new(capabilities)
     }
 }
