+++ query-engine/connectors/query-connector/src/interface/mod.rs
@@ -1,7 +1,4 @@
-mod capabilities;
 mod dispatch;
-
-pub use capabilities::*;
 pub use dispatch::*;
 
 use crate::{Filter, QueryArguments, WriteArgs};
@@ -14,9 +11,6 @@ use prisma_value::PrismaValue;
 pub trait Connector {
     /// Returns a connection to a data source.
     async fn get_connection(&self) -> crate::Result<Box<dyn Connection>>;
-
-    /// Returns supported capabilities of the connector.
-    fn capabilities(&self) -> ConnectorCapabilities;
 
     /// Returns name of the connector.
     fn name(&self) -> String;
