+++ query-engine/connectors/sql-query-connector/src/database/sqlite.rs
@@ -1,7 +1,6 @@
 use super::connection::SqlConnection;
 use crate::{FromSource, SqlError};
 use async_trait::async_trait;
-use connector::ConnectorCapabilities;
 use connector_interface::{
     self as connector,
     error::{ConnectorError, ErrorKind},
@@ -92,10 +91,6 @@ impl Connector for Sqlite {
             Ok(Box::new(conn) as Box<dyn Connection>)
         })
         .await
-    }
-
-    fn capabilities(&self) -> ConnectorCapabilities {
-        ConnectorCapabilities::default()
     }
 
     fn name(&self) -> String {
