+++ query-engine/core/Cargo.toml
@@ -18,6 +18,7 @@ once_cell = "1.3"
 petgraph = "0.4"
 prisma-inflector = {path = "../../libs/prisma-inflector"}
 prisma-models = {path = "../../libs/prisma-models"}
+datamodel-connector = {path = "../../libs/datamodel/connectors/datamodel-connector"}
 prisma-value = {path = "../../libs/prisma-value"}
 rust_decimal = {git = "https://github.com/pimeys/rust-decimal", branch = "pgbouncer-mode"}
 serde = {version = "1", features = ["derive"]}
