+++ query-engine/core/src/schema_builder/input_types/field_filter_types.rs
@@ -1,5 +1,5 @@
 use super::*;
-use connector::ConnectorCapability;
+use datamodel_connector::ConnectorCapability;
 use prisma_models::{dml::DefaultValue, PrismaValue};
 
 /// Builds filter type for the given model field.
