+++ query-engine/query-engine/src/tests/dmmf.rs
@@ -1,4 +1,4 @@
-use connector::ConnectorCapabilities;
+use datamodel_connector::ConnectorCapabilities;
 use prisma_models::DatamodelConverter;
 use query_core::{schema_builder, BuildMode, QuerySchema};
 use serial_test::serial;
@@ -129,7 +129,7 @@ fn get_query_schema(datamodel_string: &str) -> (QuerySchema, datamodel::dml::Dat
     let internal_ref = internal_dm_template.build("db".to_owned());
 
     (
-        schema_builder::build(internal_ref, BuildMode::Modern, false, ConnectorCapabilities::default()),
+        schema_builder::build(internal_ref, BuildMode::Modern, false, ConnectorCapabilities::empty()),
         dm,
     )
 }
